requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cohere==5.13.4
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
dacite==1.9.2
diskcache==5.6.3
distro==1.9.0
fastapi==0.115.6
fastavro==1.10.0
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.12.0
geopandas==1.0.1
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.37.0
groq==0.13.1
h11==0.14.0
htmlmin==0.1.12
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.0
idna==3.10
ImageHash==4.3.1
Jinja2==3.1.5
jiter==0.8.2
joblib==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kaleido==0.2.1
kiwisolver==1.4.8
langchain-core==0.3.28
langchain-groq==0.2.2
langsmith==0.2.6
lida==0.0.14
llmx==0.0.21a0
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-venn==1.1.1
mdurl==0.1.2
mizani==0.13.1
multimethod==1.12
narwhals==1.19.1
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
openai==1.58.1
orjson==3.10.12
packaging==24.2
pandas==2.2.3
parameterized==0.9.0
patsy==1.0.1
phik==0.12.4
pillow==11.0.0
plotly==5.24.1
plotnine==0.14.4
protobuf==5.29.2
puremagic==1.29
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.4
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.18.0
pyogrio==0.10.0
pyparsing==3.2.0
pyproj==3.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyWavelets==1.8.0
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
scipy==1.14.1
seaborn==0.13.2
shapely==2.0.6
shellingham==1.5.4
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.41.3
statsmodels==0.14.4
streamlit==1.41.1
tenacity==9.0.0
tiktoken==0.8.0
tokenizers==0.21.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typeguard==4.4.2
typer==0.15.1
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
visions==0.8.1
watchdog==6.0.0
wordcloud==1.9.4
ydata-profiling==4.16.1
